feat: jwt 인증 도입2

+++ controller/CartController.js
@@ -66,15 +66,29 @@ const getCartItems = (req, res) => {
 const removeCartItem = (req, res) => {
     let cartItem_id = req.params.id;
 
-    let sql = `DELETE FROM cartItems WHERE id = ?`;
-    conn.query(sql, cartItem_id, (err, results) => {
-        if (err) {
-            console.log(err);
-            return res.status(StatusCodes.BAD_REQUEST).end();
-        }
+    let authorization = authUtils.ensureAuthorization(req, res);
 
-        return res.status(StatusCodes.OK).json(results);
-    });
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            message: "로그인 세션이 만료되었습니다. 다시 로그인하세요.",
+        });
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            message: "잘못된 토큰입니다.",
+        });
+    }else {
+        
+            let sql = `DELETE FROM cartItems WHERE id = ?`;
+            conn.query(sql, cartItem_id, (err, results) => {
+                if (err) {
+                    console.log(err);
+                    return res.status(StatusCodes.BAD_REQUEST).end();
+                }
+        
+                return res.status(StatusCodes.OK).json(results);
+            });
+
+    }
 };
 
 export { addToCart, getCartItems, removeCartItem };
